+++ climpred/tests/test_PredictionEnsemble.py
@@ -172,6 +172,17 @@ def test_hindcastEnsemble_plus_broadcast(hind_ds_initialized_3d, operator):
     assert_PredictionEnsemble(he2, he3)
 
 
+def test_hindcastEnsemble_operator_different_datasets(
+    hind_ds_initialized_1d, observations_ds_1d
+):
+    """Test that HindcastEnsemble math operator (+-*/) on different datasets attached to HindcastEnsemble."""
+    he = HindcastEnsemble(hind_ds_initialized_1d)
+    he = he.add_observations(observations_ds_1d)
+    he2 = HindcastEnsemble(hind_ds_initialized_1d)
+    assert not (he2 - he).equals(he2)
+    assert not (he - he2).equals(he)
+
+
 def test_HindcastEnsemble_area_weighted_mean(hind_ds_initialized_3d):
     """Test area weighted mean HindcastEnsemble."""
     he = HindcastEnsemble(hind_ds_initialized_3d)
@@ -434,3 +445,35 @@ def test_PredictionEnsemble_cf(pe):
     for k, v in CF_LONG_NAMES.items():
         if k in pe.coords:
             assert len(pe.get_initialized().coords[k].attrs["description"]) > 5
+
+
+def test_warn_if_chunked_along_init_member_time(
+    hindcast_hist_obs_1d, perfectModelEnsemble_initialized_control
+):
+    """Test that _warn_if_chunked_along_init_member_time warns."""
+    he = hindcast_hist_obs_1d
+    with pytest.warns(UserWarning, match="is chunked along dimensions"):
+        he_chunked = HindcastEnsemble(
+            he.get_initialized().chunk({"init": 10})
+        ).add_observations(he.get_observations())
+        with pytest.raises(
+            ValueError, match="pass ``allow_rechunk=True`` in ``dask_gufunc_kwargs``"
+        ):
+            he_chunked.verify(
+                metric="rmse", dim="init", comparison="e2o", alignment="same_inits"
+            )
+
+    with pytest.warns(UserWarning, match="is chunked along dimensions"):
+        he_chunked = HindcastEnsemble(he.get_initialized()).add_observations(
+            he.get_observations().chunk({"time": 10})
+        )
+        with pytest.raises(
+            ValueError, match="pass ``allow_rechunk=True`` in ``dask_gufunc_kwargs``"
+        ):
+            he_chunked.verify(
+                metric="rmse", dim="init", comparison="e2o", alignment="same_inits"
+            )
+
+    pm = perfectModelEnsemble_initialized_control
+    with pytest.warns(UserWarning, match="is chunked along dimensions"):
+        PerfectModelEnsemble(pm.get_initialized().chunk({"init": 10}))